load lib

+++ src/com/andr0day/andrinsight/MyApplication.java
@@ -6,12 +6,16 @@ import com.andr0day.andrinsight.common.FileUtils;
 import com.andr0day.andrinsight.common.RootUtil;
 import com.andr0day.andrinsight.common.XposedConstant;
 
+import java.io.File;
+
 public class MyApplication extends Application {
 
 
     public void onCreate() {
         super.onCreate();
-        FileUtils.copyAssetsToFiles(this, XposedConstant.SO_NAME);
+        if (!new File(XposedConstant.SO_FULL_PATH).exists()) {
+            FileUtils.copyAssetsToFiles(this, XposedConstant.SO_NAME);
+        }
         RootUtil.safeExecStr("chmod 777 " + XposedConstant.SO_FULL_PATH);
         startService(new Intent(this, FileModifyService.class));
     }
